test: update existing unit test cases

+++ src/main/java/solutions/binarysearch/MinDays_1482.java
@@ -34,7 +34,7 @@ public class MinDays_1482 {
         int n = bloomDay.length;
 
         /* Corner case */
-        if (n < m * k) {
+        if ((long) n < (long) m * k) {
             return -1;
         }
